feat(manager-react-components): migrate links component to ods 19
ref: #MANAGER-18712
