+++ packages/manager-react-components/src/components/Links/index.ts
@@ -1,0 +1,3 @@
+export { Links } from './Links.component';
+export type { LinksProps } from './Links.props';
+export { LinkType } from './Links.props';
