+++ packages/manager-react-components/src/components/typography/index.ts
@@ -1,3 +1,2 @@
 export * from '../navigation/card/card.component';
 export * from './title/title.component';
-export * from './links/links.component';
