+++ packages/manager-wiki/stories/manager-react-components/typography/Links/links.stories.tsx
@@ -4,7 +4,7 @@ import { Meta } from '@storybook/react';
 import { Links, LinksProps, LinkType } from '@ovh-ux/manager-react-components';
 
 const backLink: LinksProps = {
-  label: 'Back to the list',
+  children: 'Back to the list',
   href: 'https://www.ovhcloud.com',
   target: '_blank',
   type: LinkType.back,
@@ -13,19 +13,19 @@ const backLink: LinksProps = {
 const nextLink: LinksProps = {
   href: 'https://www.ovhcloud.com/',
   target: '_blank',
-  label: 'Next Page',
+  children: 'Next Page',
   type: LinkType.next,
 };
 
 const externalLink: LinksProps = {
   href: 'https://www.ovhcloud.com/',
   target: '_blank',
-  label: 'External Page',
+  children: 'External Page',
   type: LinkType.external,
 };
 
 const meta: Meta<LinksProps> = {
-  title: 'Manager React Components/Typography/Links',
+  title: 'Manager React Components/Components/Links',
   decorators: [(story) => <div>{story()}</div>],
   argTypes: {},
   args: backLink,
